pyproject.toml
[project]
name = "data-curator"
version = "0.1.0"
description = ""
authors = [
    {name = "Codex",email = "codex@openai.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "PyYAML>=6.0",
    "pytest>=7.4",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
packages = [{include = "data_curator_app"}]
